# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ecrossflow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/settings:
    put:
      operationId: updateSettings
      tags: [users]
      summary: Update user settings (theme, language, currency)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # BOARDS ENDPOINTS
  /boards:
    get:
      operationId: getBoards
      tags: [boards]
      summary: Get all boards configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                type: object
                properties:
                  boards:
                    type: array
                    items:
                      $ref: "#/components/schemas/Board"
                required:
                  - boards
  /boards/{boardId}/instance:
    get:
      operationId: getBoardInstance
      tags: [boards]
      summary: Get active instance for a board
      security:
        - bearerAuth: []
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active board instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardInstance"
        "404":
          description: No active instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /boards/{boardId}/pay:
    post:
      operationId: payBoard
      tags: [boards]
      summary: Pay to join a board
      security:
        - bearerAuth: []
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardPaymentResponse"
        "400":
          description: Insufficient funds or ineligible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /boards/my-status:
    get:
      operationId: getMyBoardStatus
      tags: [boards]
      summary: Get user's status across all boards
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Board statuses
          content:
            application/json:
              schema:
                type: object
                properties:
                  statuses:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserBoardStatus"
                required:
                  - statuses

  # WALLET ENDPOINTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        accountNumber:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        referralCode:
          type: string
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED]
        role:
          type: string
          enum: [USER, ADMIN]
        preferredLanguage:
          type: string
        preferredCurrency:
          type: string
        preferredTheme:
          type: string
        currentBoard:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - username
        - email
        - referralCode
        - status
        - role
        - preferredLanguage
        - preferredCurrency
        - preferredTheme
        - createdAt
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    UpdateSettingsRequest:
      type: object
      properties:
        preferredLanguage:
          type: string
        preferredCurrency:
          type: string
        preferredTheme:
          type: string

    # BOARD SCHEMAS
    Board:
      type: object
      properties:
        id:
          type: string
        rankOrder:
          type: integer
        entryFee:
          type: number
        multiplier:
          type: integer
        totalGain:
          type: number
        nextBoardDeduction:
          type: number
        withdrawable:
          type: number
        description:
          type: string
          nullable: true
        colorTheme:
          type: string
          nullable: true
      required:
        - id
        - rankOrder
        - entryFee
        - multiplier
        - totalGain
        - nextBoardDeduction
        - withdrawable
    BoardParticipant:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [STARTER, CHALLENGER, LEADER, RANKER]
        position:
          type: integer
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - userId
        - username
        - role
    BoardInstance:
      type: object
      properties:
        id:
          type: string
        boardId:
          type: string
        instanceNumber:
          type: integer
        ranker:
          $ref: "#/components/schemas/BoardParticipant"
        status:
          type: string
          enum: [WAITING, ACTIVE, COMPLETED]
        slotsFilled:
          type: integer
        totalCollected:
          type: number
        participants:
          type: array
          items:
            $ref: "#/components/schemas/BoardParticipant"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - boardId
        - instanceNumber
        - ranker
        - status
        - slotsFilled
        - totalCollected
        - participants
        - createdAt
    BoardPaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        boardCompleted:
          type: boolean
        newRole:
          type: string
          nullable: true
      required:
        - success
        - message
        - boardCompleted
    UserBoardStatus:
      type: object
      properties:
        boardId:
          type: string
        role:
          type: string
          nullable: true
        instanceId:
          type: string
          nullable: true
        amountPaid:
          type: number
          nullable: true
        joinedAt:
          type: string
          format: date-time
          nullable: true
        completed:
          type: boolean
      required:
        - boardId
        - completed

    # WALLET SCHEMAS
...