# lib/api-zod/src/generated/types/board.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ecrossflow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Board {
  id: string;
  rankOrder: number;
  entryFee: number;
  multiplier: number;
  totalGain: number;
  nextBoardDeduction: number;
  withdrawable: number;
  description?: string | null;
  colorTheme?: string | null;
}